% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksea_batchKinases.R
\name{ksea_batchKinases}
\alias{ksea_batchKinases}
\title{Batch KSEA for all kinase regulons}
\usage{
ksea_batchKinases(ranking, logvalues, regulons, trial = 1000)
}
\arguments{
\item{ranking}{vector with all quantified positions}

\item{logvalues}{vector with all quantifications}

\item{regulons}{list containing vectors with substrates of each kinase. Substrates 
should be encoded with the same id as the \code{ranking} vector.}

\item{trial}{Number of permutations used for the empirical p-value generation.}
}
\value{
A vector of p-values containing all KSEA p-values for all
}
\description{
Runs KSEA in batch for all Kinase-Substrates
}
\details{
This functions allows to run the KSEA in a given dataset for all provided kinases.
The function makes use of all available cores in the computer (using \code{parallel})
package.
}
\author{
David Ochoa
}
